add a blur timeout so form button clicks work
if the blur event is triggered while a user is clicking on a form submission button, the blur callback fires however the button click is never picked up so the form is never submitted. This patchset would make the blur callback occur after 500ms. This value is arbitrary as it just needs to be long enough to allow the button to fire its own event. It's possible some other timeout value would be more correct.